move rate counter to commons
